+++ core/src/main/java/io/questdb/cairo/CairoConfiguration.java
@@ -160,7 +160,7 @@ public interface CairoConfiguration {
 
     int getParallelIndexThreshold();
 
-    long getPartitionPurgeListCapacity();
+    int getPartitionPurgeListCapacity();
 
     default Rnd getRandom() {
         Rnd rnd = RANDOM.get();
