+++ core/src/main/java/io/questdb/cairo/DefaultCairoConfiguration.java
@@ -329,7 +329,7 @@ public class DefaultCairoConfiguration implements CairoConfiguration {
     }
 
     @Override
-    public long getPartitionPurgeListCapacity() {
+    public int getPartitionPurgeListCapacity() {
         return 64;
     }
